update source location on japan module
